+++ src/NativeScript/Calling/FFICall.h
@@ -134,8 +134,8 @@ public:
         auto scope = DECLARE_THROW_SCOPE(vm);
 
         if (!this->_argumentCountValidator(this, execState)) {
-            WTF::String message = WTF::String::format("Actual arguments count: \"%lu\". Expected: \"%lu\". ", execState->argumentCount(), this->parametersCount());
-            throwException(execState, scope, JSC::createError(execState, message));
+            WTF::String message = WTF::String::format("Actual arguments count: \"%lu\". Expected: \"%lu\".", execState->argumentCount(), this->parametersCount());
+            throwException(execState, scope, JSC::createError(execState, message, defaultSourceAppender));
             return;
         }